core/pagination.py
from rest_framework.pagination import PageNumberPagination


class StandardPagination(PageNumberPagination):
    """
    Default pagination for list endpoints.
    """
    page_size = 10
    page_size_query_param = 'page_size'
    max_page_size = 50
