requirements.txt
django>=4.2,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
gunicorn>=21.0
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
PyYAML>=6.0
